IMPORANT
the material 1 -3 is a button without this code you have to press the button for it to register that you selected the color now it does it when you enter the dropdown. Code doesnt need to have a button for the material color anymore but idk if it will break if i change it.